web_api/pyproject.toml
[tool.poetry]
name = "timelink"
version = "0.1.0"
description = ""
authors = ["ChengTze-Wu <yaoop3050777@gmail.com>"]
license = "LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.8.0"
alembic = "^1.12.1"
requests = "^2.31.0"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
faker = "^21.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["web_app"]